Add MACs field in meshviewer-ffrgb.
When you traceroute through the Batman network, you see kinds of MAC addresses.
So when we have a MAC address, we should be able to detect the corresponding device.

In Leipzig, we use distinct Batman networks and one router can be attached to several
of those networks.

That patch is not enough to support the case "serveral Batman networks" as we would
have need to join all devices. But this is one should make the discussion open,
how we can support routers that are connected to multiple networks.